SWDEV-492049 - Remove the handle of Phy Mem from Memobj for VMM
The physical handle is not needed since the Phy mem obj
will be saved in virtual mem obj during mapping. And if
the phy mem handle will tracked by Memobj, then the whole
address ranged from handle to handle + size will be tracked.
This fixed the potential issue that the clr find the wrong
Memobj if we do not remove the phy mem handle from Memobj.

Change-Id: Ia442adb8e5b1cbed2d15e8fe9221ae8a0db82d8c